+++ src/main/java/redis/clients/jedis/Jedis.java
@@ -3060,7 +3060,7 @@ public class Jedis extends BinaryJedis implements JedisCommands,
 	return client.getStatusCodeReply();
     }
 
-    public Long pexpire(final String key, final int milliseconds) {
+    public Long pexpire(final String key, final long milliseconds) {
 	checkIsInMulti();
 	client.pexpire(key, milliseconds);
 	return client.getIntegerReply();
